added more admin functionality as well as some methods for removing expired tokens and refactored some code

+++ src/main/java/com/nicolaspaiva/finance_vault/expenses/dto/MonthlyResumeDto.java
@@ -17,10 +17,10 @@ import lombok.NoArgsConstructor;
 @Builder
 public class MonthlyResumeDto {
 
-    private float totalExpenses;
+    private double totalExpenses;
 
-    private float totalDeposits;
+    private double totalDeposits;
 
-    private float monthlyBalance;
+    private double monthlyBalance;
 
 }
